Basic Component Directory restructure for Game(s)
Next step is to actually refactor the code into more distinct components
