+++ jsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
       "target": "ES5",
-      "module": "es6",
+      "module": "commonjs",
       "outDir": "build",
       "allowSyntheticDefaultImports": true
     },
