+++ test/components/Game/GameMenu.test.js
@@ -1,11 +1,11 @@
-import { expect } from '../test-setup';
+import { expect } from '../../test-setup';
 import { findRenderedComponentWithType } from 'react-dom/test-utils';
 import { Sidebar } from 'semantic-ui-react';
 
-import { wrapContainer } from '../support/container-helper';
-import GameDummy from '../dummies/game';
+import { wrapContainer } from '../../support/container-helper';
+import GameDummy from '../../dummies/game';
 
-import GameMenu from '../../src/components/GameMenu';
+import GameMenu from '../../../src/components/Game/GameMenu';
 
 describe('(Component) GameMenu', () => {
   context('before game has started', () => {
